Remove `loadSnapshot` from `StateSnapshotable`
I think that an important invariant in DDD for entities is that their
identity should remain stable: it should never change.

Despite the fact that the identity of `Aggregate1` is immutable and that
`loadSnaphot` checks that the snapshot's identity is the same as the
aggregate's identity, I find it somewhat misleading. It is also
unnecessary, because the `fromSnapshot` factory method is all we need.